% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_combn.R
\name{append_combn}
\alias{append_combn}
\title{Append a combination of head and tail elements to x}
\usage{
append_combn(x, head = NULL, tail = NULL)
}
\arguments{
\item{x}{A vector of values you wish to append to; will be coerced to \code{character} if not.}

\item{head}{(optional) A vector of values you wish to append to the head of x.}

\item{tail}{(optional) A vector of values you wish to append to the tail of x.}
}
\value{
A \code{list} of length \emph{x}, with elements the length of the cartesian product of
\code{length(head)} and \code{length(tail)}.
}
\description{
Create a combination of patterned outputs of x based on head and/or tail elements
}
\examples{
x <- replicate(10, paste(sample(letters, 3), collapse = ""))
append_combn(x, head = ":", tail = "-")
append_combn(x, head = c(":", "|"), tail = "-")
append_combn(x, head = "_A_")
append_combn(x, head = paste0(LETTERS[1:10], "_"), tail = c("-", "_"))
}
