% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_rand.R
\name{cas_rand}
\alias{cas_rand}
\title{Generate random validly formatted CAS RN}
\usage{
cas_rand(n)
}
\arguments{
\item{n}{integer. How many random CAS do you wish to generate? \code{numeric} or \code{character}
input will be coerced to integer, if possible. Negative (valid) input will be transformed to
positive via \code{abs}}
}
\value{
A vector of random CAS RNs of length \emph{n}.
}
\description{
Make random CAS RN of valid length format, without checksum consideration
}
\details{
Generate \emph{n} random CAS with valid CAS digit length, per section and overall. Random applies to
all three sections of a CAS, and for the first section, lengths will be randomly betwen \code{2:7}.

This function does not perform any additional validation, which means duplicate CAS are possible.
More importantly, there is no last (checksum) digit validation.
}
\note{
This is mainly useful as a testing tool for \code{\link{cas_checkSum}}. See examples.
}
\examples{
set.seed(1)
random_CASRN <- cas_rand(1E4L)
length(unique(random_CASRN))
table(nchar(random_CASRN)-2) #-2 for two dashes
#using cas_detect and cas_checkSum
table(cas_detect(unique(random_CASRN)))
table(cas_checkSum(unique(random_CASRN)))
}

