% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_issn.R
\name{check_issn}
\alias{check_issn}
\title{Check ISSN validity}
\usage{
check_issn(x)
}
\arguments{
\item{x}{A \code{character} vector of ISSNs to check}
}
\value{
A \code{list} of length 2 containing a data.frame of results and any ISSNs
that could not be checked due to invalid format; if no inputs are invalid, then
\code{NULL} for the latter.
}
\description{
Calculate the check digit of ISSNs and compare to current digit
}
\details{
Calculates the check digit of one or more input fundamentally-formatted ISSNs.
Fundamentally formatted refers to the input after whitespace-stripping, and only
keeping alphanumeric values Additionally, the final character checksum is performed
on a case-normalized basis.

This means that input ISSNs can be quite ill-formatted and still be evaluated for
validity, and test as valid. See example.
}
\examples{
# Note that dashes and whitespaces are
test_issns <- c("10543589", "00657743", "0091679X",
                "1054-3589", "0065-7743", "0091-679X",
                "1054--3589", " 0065-7743", "0091--679x-",
                "0091679xx", "0091x79x", "0091X679")
check_issn(test_issns)
}
