% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlProgress.R
\name{curlProgress}
\alias{curlProgress}
\title{curlProgress}
\usage{
curlProgress(url, fname, mode = "wb", width_fx = 0.8)
}
\arguments{
\item{url}{A remote URL}

\item{fname}{The local destination, including the file name and extension}

\item{mode}{Mode, as in RCurl::CFILE. Read, Write; Text, Binary. Defaults to 'wb' (write binary)}

\item{width_fx}{(Optional) The console output message width; defaults to \code{0.8}}
}
\value{
The requested asset
}
\description{
Download a binary file from remote to local, with progress monitor
}
\details{
This is unabashedly taken from https://stackoverflow.com/a/21961094, and all credit is due. Minor
additional input checks.
}
\examples{
\dontrun{
# ftp example
curlProgress(
  url = "ftp://user:pwd@.xyz.com/dir/myfile.zip",
  fname = "./local_dir/myfile.zip"
)
}
}
\seealso{
\code{\link{ftp_getBinPars}} for creating FTP parameters to pass to this function.
}
