% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_tbl_names.R
\name{fix_tbl_names}
\alias{fix_tbl_names}
\title{Clean column names from a data.frame or chr vector}
\usage{
fix_tbl_names(x, lower = TRUE)
}
\arguments{
\item{x}{A character vector or a data.frame. If the latter, will extract names}

\item{lower}{Lowercase the output? Defaults to \code{TRUE}}
}
\value{
A character vector of cleaned-up names. Any duplicated names are handled by \code{make.unique()}.
}
\description{
Remove non-alphanum, replace whitespace with uscore, optionally lower, dedupe
}
\details{
This is a simple utility that replaces any non-alphanumeric characters in \emph{x} with an
underscore, ensuring that runs of underscores > 1 are truncated to length 1, stripping any
trailing underscores, and optionally lowercasing the output (default \code{TRUE}).

Furthermore, any duplicated values after the above are handled via \code{make.unique()}
}
\examples{
x <- c("This is (a dirty) name  with ", "A", "a")
fix_tbl_names(x)
fix_tbl_names(x, lower = FALSE)

xdf <- data.frame(1:10, LETTERS[1:10], letters[1:10])
names(xdf) <- x
fix_tbl_names(xdf)
}
