% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_makeParams.R
\name{ftp_makeParams}
\alias{ftp_makeParams}
\title{Construct params to interact with an FTP endpoint}
\usage{
ftp_makeParams(host, protocol = "ftp", user = NULL, pwd = NULL, extras = NULL)
}
\arguments{
\item{host}{(required) The host address string. Should take the form of \code{ftp.your_host.com}}

\item{protocol}{(optional) The protocol to use; defaults to \code{ftp}}

\item{user}{(optional) User login name, if host access requires authentication. If provided, \emph{pwd}
must also be a valid string}

\item{pwd}{(optional) User password, if host access requires authentication. If provided, \emph{user}
must also be a valid string}

\item{extras}{(optional) Additional arguments to pass to output. If provided, must be a named
vector.}
}
\value{
If \emph{extras} are not supplied, a named list of length 2, containing the \code{url} and \code{userpwd}.
If either \emph{user} or \emph{pwd} are missing OR are blank, the second element will be \code{NULL}.

If \emph{extras} are supplied, a named list of length 2 plus \code{length(extras)}.
}
\description{
Create url strings and, optionally, authentication parameters for FTP endpoint interactions
}
\details{
This is a simple helper function that is most useful for constructing argument-lists to pass to e.g.
\code{\link[RCurl]{getURL}}; the outputs are named with this use case in mind. See examples.

This function normalizes common variants of input formats for \emph{host} and \emph{protocol}. See examples.
}
\examples{
# with authentication
ftp_makeParams(host = "ftp.123.com", user = "myself", pwd = "my_pwd")

# NULL or blank for either of user or pwd will result in NULL userpwd
ftp_makeParams(host = "ftp.123.com", user = "myself", pwd = " ")

# host and protocol input normalization
x0 <- ftp_makeParams(host = "ftp.123.com", protocol = "ftp" )
x1 <- ftp_makeParams(host = "ftp://ftp.123.com", protocol = "ftp")
x2 <- ftp_makeParams(host = "ftp://ftp.123.com", protocol = "ftp://")
x3 <- ftp_makeParams(host = "ftp://ftp.123.com")

all(sapply(list(x1, x2, x3), function(f) identical(x0, f)))

# If you wanted to get a listing of all files:
\dontrun{
library(RCurl)
my_params <- ftp_makeParams(host = "ftp.123.com", user = "myself", pwd = "my_pwd")
my_params$ftplistonly <- TRUE
do.call(getURL, my_params)
}
}
\seealso{
Other ftp_functions: 
\code{\link{ftp_getBinPars}()},
\code{\link{ftp_listDir}()},
\code{\link{ftp_parseMeta}()}
}
\concept{ftp_functions}
