% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_parseMeta.R
\name{ftp_parseMeta}
\alias{ftp_parseMeta}
\title{Parse ftp return from getURL call with no options}
\usage{
ftp_parseMeta(ret)
}
\arguments{
\item{ret}{(required) A character vector; the return from a call to
\code{getURL} with no options}
}
\value{
A \code{data.table}
}
\description{
Parse ftp return from getURL call with no options
}
\examples{
# TBD
}
\seealso{
Other ftp_functions: \code{\link{ftp_listDir}},
  \code{\link{ftp_makeParams}}
}
