% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chr_logi.R
\name{make_chr_logi}
\alias{make_chr_logi}
\title{Coerce one or values in a character vector to logical TRUE}
\usage{
make_chr_logi(x, val_as_TRUE = NULL, preserve_NA = TRUE, NA_as_what = NA)
}
\arguments{
\item{x}{Input character vector to coerce to logical}

\item{val_as_TRUE}{A character vector containing the values to coerce to TRUE. Including \code{NA} will
override both \emph{preserve_NA} and \emph{NA_as_what}}

\item{preserve_NA}{Should \code{NA} be preserved? Acts as a safety flag, and \emph{NA_as_what} is ignored
unless this is set to \code{FALSE}; defaults to \code{TRUE}}

\item{NA_as_what}{How should \code{NA} be handled? One of \code{TRUE, FALSE, NA}; defaults to \code{NA}.
Ignored if \emph{preserve_NA} is \code{TRUE}}
}
\value{
A logical vector of length \emph{x}
}
\description{
Transform a character into logical with denoted values as TRUE, with NA handling options
}
\details{
This function is most useful for its explicit handling of \code{NA}. By default, \code{NA} is
passed-through. If \code{NA} is included within \emph{val_as_TRUE}, then all \code{NA} will be
set to \code{TRUE}. Alternatively, if \emph{preserve_NA} is set to \code{FALSE} and \emph{NA_as_what}
is set to either \code{TRUE} or \code{FALSE}, \code{NA} will be set as requested.

Non-character inputs to either \emph{x} or \emph{val_as_TRUE} will raise an error. Furthermore,
if \emph{x} is all \code{NA}, an error is raised.
}
\examples{
x <- c("yes", "YES", NA, "no", "NO", "NA", "0")
make_chr_logi(x, c("yes", "YES"))
make_chr_logi(x, c("yes", "YES", NA)) # overrides preserve_NA or NA_as_what
}
