% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_trailing_slash.R
\name{norm_trailing_slash}
\alias{norm_trailing_slash}
\title{Handle trailing (and optionally, leading) slashes}
\usage{
norm_trailing_slash(
  x,
  how = c("remove", "replace"),
  replace_with = c("forward", "back"),
  strip_leading = FALSE,
  view_res = FALSE
)
}
\arguments{
\item{x}{A character vector}

\item{how}{What kind of normalization should be performed? One of \code{remove,replace}. Defaults to \code{remove}}

\item{replace_with}{If \code{how = replace}, should we replace with a forward or backslash? One of \code{forward,back}.
Defaults to \code{forward}}

\item{strip_leading}{Also strip the first leading slash, if present? Defaults to \code{FALSE}}

\item{view_res}{Output the structure the first six rows of the input and output? Defaults to \code{FALSE}}
}
\value{
A character vector of length \emph{x}
}
\description{
Remove, or remove and insert a single trailing forward or back slash
}
\details{
\emph{Normalization} is defined as taking the input and stripping a single trailing slash, if one is present. This
is performed whether removing only, or replacing; hence, this function will remove a single trailing slash and stop,
or also append a single trailing slash (and if so, the type of slash as specified by \emph{replace_with}).There is
nothing fancy about this function; it is a useful helper for e.g. creating directory strings
}
\note{
If \code{view_res = TRUE}, the first six rows of the input and output are printed to the console via \code{dput()}
(with default args). This may be useful to verify that the literal output, sans escapes, are as expected by e.g.
calling \code{View()} on the deparsed output.
}
\examples{
norm_trailing_slash(x = c("/dir1\\\\", "/dir1/"), how = "remove", view_res = TRUE)
norm_trailing_slash(x = c("/dir1\\\\", "/dir1/"), how = "replace",
  replace_with = "forward", view_res = TRUE)
}
