% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sevenzip_parseinfo.R
\name{sevenzip_parseinfo}
\alias{sevenzip_parseinfo}
\title{Extract key metadata from common 7-zip archives}
\usage{
sevenzip_parseinfo(x)
}
\arguments{
\item{x}{A valid return from a (captured) call to 7-zip with an
action of \code{l} (\code{list})}
}
\value{
A \code{data.frame}
}
\description{
Extract key metadata from common 7-zip archives
}
\examples{
#TBD
}
\seealso{
Other sevenzip functions: \code{\link{sevenzip_checkpath}},
  \code{\link{sevenzip_listcont}},
  \code{\link{sevenzip_makecmd}}
}
