% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fill.R
\name{split_fill}
\alias{split_fill}
\title{Split and fill a chr vector}
\usage{
split_fill(dat, targ, split_on, IDcol, rebind = FALSE, keep.targ = FALSE)
}
\arguments{
\item{dat}{a data.table}

\item{targ}{chr; vector of length 1 denoting column that contains data to be split}

\item{split_on}{chr; what pattern should be used to perform the split?}

\item{IDcol}{chr; vector of length 1 denoting the column in \emph{dat} containing the ID to be used for melting}

\item{rebind}{logical; should the original columns be appended back to the output? Defaults to \code{FALSE}}

\item{keep.targ}{logical; only used if \emph{rebind} = \code{TRUE}; drop the column that was split on?}
}
\value{
A melted data.table using \emph{IDcol} as \code{id.var} for \link[data.table]{melt.data.table},
with \emph{targ} splitted by \emph{split_on}.

If \code{rebind == TRUE}, will also return the original columns, with a single \emph{IDcol} as denoted
in input. This is performed via a \code{data.table} ad-hoc join, using \emph{IDcol} in \code{j}. The input
\emph{targ} column will be returned as well, if \emph{keep.targ} is \code{TRUE}.
}
\description{
Split a chr vector based on sep, return melted DT by ID
}
\details{
This is a convenience-convenience (not a typo) wrapper around \code{data.table::tstrsplit}, taking advantage
of the performance of \code{data.table::Ctranspose}, and adding faculties to melt and rejoin selectively.
}
\note{
\emph{targ} currently is limited to a vector of length 1, as is \emph{IDcol}. This is likely to change in the future, to
make this function more flexible and consistent with the capabilities of \code{melt.data.table}.
}
\examples{
set.seed(1)
dt <- data.table(
    ID = 1:10,
    targ = sapply(1:10, function(f)
        paste(sample(LETTERS, sample(1:5)), collapse = "|"))
)
head(split_fill(dat = dt, targ = "targ", split_on = "\\\\|", IDcol = "ID"))

#Demonstrating rebind
dt[, targ_additional := targ]
head(split_fill(dat = dt, targ = "targ", split_on = "\\\\|", IDcol = "ID", rebind = TRUE))

}

