% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_listDir.R
\name{ftp_listDir}
\alias{ftp_listDir}
\title{List files and/or directories within an ftp endpoint}
\usage{
ftp_listDir(arglist, ncharlim = 3L)
}
\arguments{
\item{arglist}{(reqiured) A list of parameters to pass to \code{\link[RCurl]{getURL}},
containing at minimum \code{url}.}

\item{ncharlim}{(optional) Number of terminal characters after a dot, beyond which a character vector
is no longer considered a file. Defaults to \code{3L}.}
}
\value{
A \code{list} of length 2, each element itself \code{list} of directories and files, respectively. If
no valid values exist for either top-level return element, then \code{NULL}.
}
\description{
For a specific ftp endpoint, list all top-level files and/or directories
}
\examples{
\dontrun{
library(RCurl)
my_arglist <- ftp_makeParams(host = "ftp.123.com", user = "myself", pwd = "my_pwd")
ftp_listDir(my_arglist)
}
}
\seealso{
Other ftp_functions: \code{\link{ftp_makeParams}}
}
