% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ifall_cols.R
\name{remove_ifall_cols}
\alias{remove_ifall_cols}
\title{Remove all columns that are all of a single value}
\usage{
remove_ifall_cols(df, ifallwhat = NULL, factor_as_chr = TRUE,
  drop_all_NA = TRUE)
}
\arguments{
\item{df}{A data.frame (or data.table)}

\item{ifallwhat}{A vector of value(s) to test for. \code{NA}, if included, will be dropped}

\item{factor_as_chr}{If \emph{ifallwhat} is of class \code{character}, should columns of
class \code{factor} be treated as character (and therefore subject to deletion as well)}

\item{drop_all_NA}{Should columns that are all NA be dropped? Defaults to \code{TRUE}}
}
\value{
A \code{data.frame}, or a \code{data.table} if this was passed to \emph{df}, with
all columns that match the requirements within \emph{ifallwhat} removed. If no columns
in \emph{df} met the criteria requested, then the input is returned without modification.
}
\description{
For a data.frame, remove all columns that satisfy a condition
}
\note{
This function will coerce the input data.frame into a data.table via
\code{\link[data.table]{as.data.table}} as an intermediate step. If a data.frame is detected
as input, it will coerce the output back to a data.frame for output.
}
\examples{
my_df <- data.frame(
A = 1:3L,
B = rep(numeric(3), 3),
C = rep("", 3),
D = rep(NA_real_, 3),
stringsAsFactors = FALSE
)

my_df_1 <- remove_ifall_cols(my_df, numeric(3))
my_df_2 <- remove_ifall_cols(my_df, numeric(3), drop_all_NA = FALSE)
remove_ifall_cols(my_df, c(""), drop_all_NA = FALSE)

remove_ifall_cols(my_df, c(1:3))
remove_ifall_cols(my_df, c(1:3), drop_all_NA = FALSE)
remove_ifall_cols(my_df, as.numeric(1:3), drop_all_NA = FALSE) # does nothing
}
